Added maximum time difference on trusting clock
If the hardware clock is off by more than the threshold (currently set
to ten years), we assume the hardware clock is broken and reset it.

This is for devices that reset the clock to a far future date when they
lose power (e.g. Acer C720 Chromebook sets the year to 2151.)